fixed flake
From looking into the zaneyos flake I'm guessing the issue is with stylix since it's the only thing in the flake using cavalier. What I'm assuming the issue is that you're using the 24.11 version of home manager and nixpkgs (hence the stable) but didn't change stylix to also use 24.11 meaning you're using stylix on the unstable branch but home-manager on the stable branch making stylix use modules that are in unstable but not stable.
https://gitlab.com/Zaney/zaneyos/-/blob/main/flake.nix?ref_type=heads#L8
Change this line in your flake.nix to ``stylix.url = "github:danth/stylix/release-24.11";`` and try to rebuild also installed mission center and local llm ai tools